#201 feat: 냉장고 탈퇴 알림 메세지 생성
