--- src/main/java/com/example/icebutler_server/alarm/controller/MainController.java
@@ -1,30 +1,0 @@
-package com.example.icebutler_server.alarm.controller;
-
-import com.example.icebutler_server.alarm.dto.RequestDTO;
-import com.example.icebutler_server.alarm.service.FirebaseCloudMessageService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.io.IOException;
-
-@RestController
-@RequiredArgsConstructor
-public class MainController {
-
-    private final FirebaseCloudMessageService firebaseCloudMessageService;
-
-    @PostMapping("/api/fcm")
-    public ResponseEntity pushMessage(@RequestBody RequestDTO requestDTO) throws IOException {
-        System.out.println(requestDTO.getTargetToken() + " "
-                +requestDTO.getTitle() + " " + requestDTO.getBody());
-
-        firebaseCloudMessageService.sendMessageTo(
-                requestDTO.getTargetToken(),
-                requestDTO.getTitle(),
-                requestDTO.getBody());
-        return ResponseEntity.ok().build();
-    }
-}
