+++ src/main/java/com/example/icebutler_server/alarm/dto/FcmMessage.java
@@ -11,20 +11,11 @@ public class FcmMessage {
     private boolean validateOnly;
     private Message message;
 
-    @Builder
-    @AllArgsConstructor
-    @Getter
-    public static class Message {
-        private Notification notification;
-        private String token;
+    public static FcmMessage makeMessage(String targetToken, String title, String body) {
+        return FcmMessage.builder()
+                .message(Message.toEntity(targetToken, title, body))
+                .validateOnly(false)
+                .build();
     }
 
-    @Builder
-    @AllArgsConstructor
-    @Getter
-    public static class Notification {
-        private String title;
-        private String body;
-        private String image;
-    }
 }
