--- src/main/java/com/example/icebutler_server/alarm/dto/RequestDTO.java
@@ -1,12 +1,0 @@
-package com.example.icebutler_server.alarm.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class RequestDTO {
-    private String targetToken;
-    private String title;
-    private String body;
-}
