+++ src/main/java/com/example/icebutler_server/alarm/service/AlarmService.java
@@ -2,7 +2,9 @@ package com.example.icebutler_server.alarm.service;
 
 import com.example.icebutler_server.user.entity.User;
 
+import java.io.IOException;
+
 public interface AlarmService {
-    void sendWithdrawalAlarm(User user, String fridgeName);
+    void sendWithdrawalAlarm(User user, String fridgeName) throws IOException;
 
 }
